If we fail to fetch registration data from ES, we should make sure fields who are also stored in the browser reflect that
For example, status is stored both locally and in ES (because we need to have access to it in Sense). If you wipe the data folder of ES, the browser version stayed and would end up being stored in ES, setting the purchase flag to true. Therefore no new registration information will be asked and thus the new cluster uuid will not be connected to that info.

Also fixes failing tests that got broken because of previous changes.

Closes elastic/x-pack#265

Original commit: elastic/x-pack@223c5debfd6b2f33ba237f1acd894d50b90717e0