Stats: Move ClusterStats to use CommonStats
The ClusterStats API collects cluster wide stats. Part of the stats reported related to indices. The collected metrics were hard coded but also partial. This PR changes to use CommonStats as a data vehicle, which means both that all metrics are reported and  that new statistics will be automatically added.