IP Filtering: Add support for having on filters on HTTP transport
In order to fix elastic/x-pack#378 a problem was revealed, that the ip filter for HTTP was
always the one for the default profile, which lead to failed tests (along
with wrong socket connections, which made the test go green irregularly).

This commit fixes the tests and allow to configure own HTTP ip filters, adding
the following settings

* shield.http.filter.enabled
* shield.http.filter.allow
* shield.http.filter.deny

If not specific settings are configured, the one of the default profile are used.

Closes elastic/x-pack#378

Original commit: elastic/x-pack@7b231deaefc0e114af3b065ff6b894be211fde94