Add the option to specify both a keystore and key password
Previously, we always used the keystore password as the key password, which does not
work if a user specifies a different password when generating the key with keytool. If the
key password is not specified, the keystore password will be used.

Closes elastic/x-pack#473

Original commit: elastic/x-pack@1c9d3cb978a3131b45d82efb5aadd083c9a98f6a