Files: Create temp files first instead of overwriting existing files
Until now the store implementations truncated the existing file and recreated it in-place.
This has problems with bigger files as the files may not yet have finished to be written,
when the WatchService tries to pick them up.

This commit fixes the problem by creating a temp file first to write the new roles or user files
and then moves it over. Also an atomic move is tried first, if the file system supports this.

Closes elastic/x-pack#477

Original commit: elastic/x-pack@ffff7b705308787f72c906bd218fb8f9a9fa0a90