[Test] invalidate any existing sessions prior to testing hostname verification
For LDAP hostname verification, we use the "default" SSLContext, which is cached in a map
and re-used. If a secure connection is established then the session is cached for use later. In
the tests, we sometimes run a test that connects without hostname verification and a SSL session
is cached. Then when the hostname verification test runs, it uses the cached session and does
not perform hostname verification causing the test to fail. This fix changes the test to always
use a new SSLContext for each test.

Closes elastic/x-pack#521

Original commit: elastic/x-pack@ba42902c65e29e3c8087cf055214db81680205e0