Bug 961 - Fix retrying of stats persistance
If the version conflict exception occurs during the
execution of the request, the exception will be wrapped
in, eventually, an ExecutionException. That is the
reason the try-catch block was not catching the exception.

In fact, we do not have to implement the retry logic
ourselves. The UpdateRequestBuilder provides functionality
to set a number of retries.


Former-commit-id: ccdce96be90eadc2dc4bb0455c721c5562bf0064

Original commit: elastic/x-pack@4a53ff0c1a6927fa5d11ea0d907fa29e2772e958