[authz] fix Indices.Globals iterator to continue past a global without indices
If a Global permission is encountered that has no indices the iterator would not continue
to the rest of the entries. This change fixes the iterator to check if a Global has indices and
if not continues to look for a Global with indices permissions.

Closes elastic/x-pack#701

Original commit: elastic/x-pack@6469053ab1576ebfbe97b2a90351587a6b6a3f8b