Correct lack of sync of normalized probability
Lack of sync in this case means that the
maxNormalizedProbability of the bucket does
not match the max normalizedProbability of
its records.

When we perform normalisation, the normalized
probability of each record gets calculated and
set. The max of those gets set on the bucket.
However, this has two problems:

  - It is possible that a records probability
  changes significantly, yet the bucket does not.
  This will lead to the record being written on
  ES while the bucket does not get updated, letting
  them out of sync.
  - The maxNormalizedProbability on the bucket should
  be calculated among the normalized probabilities of
  records with big change and the pre-normalized
  probabilities of records that don't get to have
  a big change.

This commit fixes the aforementioned issues.


Former-commit-id: 3e1a7f3aa5b98b3f4752f5359c37b13c161df2c0

Original commit: elastic/x-pack@3f2037110498ae522d9b3f6a83cae6f32fc97934