Netty: reduce logging for a close request during handshake
Reduces the amount of logging on both HTTP and Transport protocols for
a channel being closed while in the middle of a handshake. This often
occurs when the client does not trust the server certificate and aborts the
handshake. Also, reduces logging on the Transport protocol for a
plain text message received on a channel that is using TLS.

Closes elastic/x-pack#771

Original commit: elastic/x-pack@0b5f0a6f280f3fbc58155e92214eac16d7427263