Fix silent failure to parse valid script condition.
`
{
  "script" : "<script>"
}
`
Should be a valid script condition. This was silently failing to parse creating invalid watches in the system that could not be executed or interacted with except via DELETE.
Added a test for this as well.

Fixes elastic/x-pack#191

Original commit: elastic/x-pack@ad4726d4f467be2dd0bd268a54a553311c47ba77