While overwriting bucket, also delete its records
In the case where a bucket is flushed and interim results
are calculated, the interim records will be stored.

When later the final results will be calculated or if
the bucket gets reset, etc., we need to delete the
records of the bucket before we persist the newest ones.


Former-commit-id: c21514e0066448d7899f97ac52c5d6f003b8d4a2

Original commit: elastic/x-pack@e779b30aff0ad418e3d8cf8c09613d208934f942