[CLI] check the number of arguments for esusers commands
Adds a check to all of the esusers commands to ensure the correct number
of arguments are found. If extra arguments are found, they are printed out
with an error message and the tool exits.

Closes elastic/x-pack#817

Original commit: elastic/x-pack@ca69a90de600f4f79c72426fb3b174cf221b58a7