[Test] fix bug in creating a byte array for tampered encryption test
The testChangingAByte method was requesting a random length byte array ranging
from 0-1000. The issue is that a byte array with length of 0 cannot be changed and
therefore this test is not valid in this case. It now requests a range of 1-1000.

Original commit: elastic/x-pack@c0d01015f061a9f3eb992c1e4ddb690c7eb6bde6